Refactor analysis modules
